These are new sound effects by John Alston
Custom recorded. POSTSCRIPT:  I initially mastered the audio for the goblin death to have an understated quality,  so that it's an informational cue, "not this HOLY COW I annihilated the crap out of this enemy" -- I like the 'smoke' effect in games like A Link to the Past or Aladdin.